Fix definition of the x87 ftag register

+++ Headers/DebugServer2/Architecture/X86/CPUState.h
@@ -59,7 +59,7 @@ struct CPUState {
     X87Register regs[8];
     uint16_t fstw;
     uint16_t fctw;
-    uint8_t ftag;
+    uint16_t ftag;
     uint32_t fiseg;
     uint32_t fioff;
     uint32_t foseg;
